# Improve compilation time and execution time for quantized matmul on ARM