--- a/iree/compiler/Codegen/LLVMCPU/KernelDispatch.cpp
+++ b/iree/compiler/Codegen/LLVMCPU/KernelDispatch.cpp
@@ -369,22 +369,25 @@ static LogicalResult setDefaultRootConfig(
       DispatchLoweringPassPipeline::CPUDefault);
 }
 
-static LogicalResult setX86SandboxRootConfig(func::FuncOp entryPointFn,
-                                             linalg::ContractionOpInterface op,
-                                             ArrayRef<int64_t> flowTileSizes,
-                                             int vectorSize) {
-  // Hardcoded tiling sizes {1, 1, ..., 8, 32, 16}.
+static LogicalResult setSandboxRootConfig(func::FuncOp entryPointFn,
+                                          linalg::ContractionOpInterface op,
+                                          ArrayRef<int64_t> flowTileSizes,
+                                          ArrayRef<int64_t> target2ndTileSizes,
+                                          int vectorSize) {
+  assert(target2ndTileSizes.size() == 3 &&
+         "the current configuration is driven by matmul which has exactly "
+         "three loops");
   // The tiling for parallel dims and reduction dims should be separated.
   SmallVector<int64_t> parallelTileSizes;
   int64_t nLoops = cast<linalg::LinalgOp>(op.getOperation()).getNumLoops();
   if (nLoops >= 3) {
     parallelTileSizes.append(nLoops - 3, 1);
-    parallelTileSizes.push_back(
-        getMaxTileSize(0, flowTileSizes[nLoops - 3], 8, vectorSize));
+    parallelTileSizes.push_back(getMaxTileSize(
+        0, flowTileSizes[nLoops - 3], target2ndTileSizes[0], vectorSize));
   }
   if (nLoops >= 2) {
-    parallelTileSizes.push_back(
-        getMaxTileSize(0, flowTileSizes[nLoops - 2], 32, vectorSize));
+    parallelTileSizes.push_back(getMaxTileSize(
+        0, flowTileSizes[nLoops - 2], target2ndTileSizes[1], vectorSize));
   }
   parallelTileSizes.push_back(0);
 
@@ -392,7 +395,8 @@ static LogicalResult setX86SandboxRootConfig(func::FuncOp entryPointFn,
   int64_t K = lhsShapedType.getShape().back();
   SmallVector<int64_t> reductionTileSizes;
   reductionTileSizes.append(nLoops - 1, 0);
-  reductionTileSizes.push_back(getMaxTileSize(0, K, 16, vectorSize));
+  reductionTileSizes.push_back(
+      getMaxTileSize(0, K, target2ndTileSizes[2], vectorSize));
 
   TileSizesListType tileSizes;
   tileSizes.emplace_back(flowTileSizes.begin(), flowTileSizes.end());
@@ -494,13 +498,22 @@ static LogicalResult setRootConfig(
     // number of loops when adding the pass to pass manager.
     // TODO(hanchung): Embed options into attributes, so we can control options
     // more heuristically.
-    return setX86SandboxRootConfig(entryPointFn, contractionOp, flowTileSizes,
-                                   vectorSize);
+    SmallVector<int64_t> tileSizes = {8, 32, 16};
+    return setSandboxRootConfig(entryPointFn, contractionOp, flowTileSizes,
+                                tileSizes, vectorSize);
   }
 
   // Fall back to ARM configurations.
-  return setARMRootConfig(entryPointFn, contractionOp, flowTileSizes,
-                          vectorSize);
+  bool isQuantized =
+      lhsShapedType.getElementType() != resShapedType.getElementType();
+  if (isQuantized) {
+    SmallVector<int64_t> tileSizes = {4, 16, 4};
+    return setSandboxRootConfig(entryPointFn, contractionOp, flowTileSizes,
+                                tileSizes, vectorSize);
+  } else {
+    return setARMRootConfig(entryPointFn, contractionOp, flowTileSizes,
+                            vectorSize);
+  }
 }
 
 /// Sets the lowering configuration for dispatch region for linalg.mmt4d root